+++ components/case-studies/case-studies-content.tsx
